refactor: 重命名模块以符合 snake_case: SSO -> sso; SSO_error -> sso_error
